refactor(settings): uses snackbar component for streaming sync
